double precision

+++ component/Camera.h
@@ -8,12 +8,12 @@
 typedef struct _CameraInfo
 {
 	//projection related variables
-	float		fov;
-	float		aspect;
-	float		near;
-	float		far;
+	double		fov;
+	double		aspect;
+	double		near;
+	double		far;
 
-	Vector2f	jitterOffset;
+	Vector2d	jitterOffset;
 }CameraInfo;
 
 class Camera : public BaseComponent
@@ -24,18 +24,18 @@ public:
 	void Update() override;
 	void OnPreRender() override;
 
-	void SetFOV(float new_fov);
-	void SetAspect(float new_aspect);
-	void SetNearPlane(float new_near_plane);
-	void SetFarPlane(float new_far_plane) { m_cameraInfo.far = new_far_plane; m_projDirty = true; }
-	void SetJitterOffset(Vector2f jitterOffset);
+	void SetFOV(double new_fov);
+	void SetAspect(double new_aspect);
+	void SetNearPlane(double new_near_plane);
+	void SetFarPlane(double new_far_plane) { m_cameraInfo.far = new_far_plane; m_projDirty = true; }
+	void SetJitterOffset(Vector2d jitterOffset);
 
 	void SetCameraInfo(const CameraInfo& info);
 
 	const CameraInfo GetCameraInfo() const { return m_cameraInfo; }
-	const float GetFovH() const { return m_fovH; }
+	const double GetFovH() const { return m_fovH; }
 
-	const Matrix4f GetVPMatrix() const { return m_vpMatrix; }
+	const Matrix4d GetVPMatrix() const { return m_vpMatrix; }
 
 	static std::shared_ptr<Camera> Create(const CameraInfo& info);
 
@@ -51,13 +51,13 @@ protected:
 
 	//Camera supplement info, no need to update by client, but to be calculated dynamically
 	//make it as member variable just for convinience
-	float		m_fovH;		//field of view in horizontal, required by some circumstances like character controller
+	double		m_fovH;		//field of view in horizontal, required by some circumstances like character controller
 
-	Matrix4f	m_vpMatrix;
+	Matrix4d	m_vpMatrix;
 
 	//float		m_viewDirty;
-	float		m_projDirty;
+	double		m_projDirty;
 
 	// Halton 2x3 for camera view frustom jittering
-	float		m_jitterPattern[16 * 2];
+	double		m_jitterPattern[16 * 2];
 };
